backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pyJWT==2.8.0
pymongo==4.6.1
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
google-generativeai==0.3.2
PyPDF2==3.0.1
pdfkit==1.0.0
Werkzeug==3.0.1
geopy==2.4.1